# docs/openapi-app.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catálogo de Filmes — API",
    "version": "6.0.0",
    "description": "OpenAPI da aplicação principal. A identidade do perfil é derivada da sessão autenticada; o backend não confia em usuario_id enviado no corpo."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Criar conta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastro criado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "409": {
            "description": "E-mail já cadastrado"
          },
          "500": {
            "description": "Falha ao enviar confirmação"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Entrar no sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado"
          },
          "401": {
            "description": "Credenciais inválidas"
          },
          "403": {
            "description": "E-mail não confirmado"
          },
          "503": {
            "description": "Auth indisponível"
          }
        }
      }
    },
...
    "/api/auth/verify-email": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Confirmar e-mail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail confirmado"
          },
          "400": {
            "description": "Token inválido"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Solicitar recuperação de senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitação processada"
          },
          "400": {
            "description": "E-mail ausente"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Redefinir senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "novaSenha"
                ],
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "novaSenha": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada"
          },
          "400": {
            "description": "Token ou senha inválidos"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
...
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 150
          },
          "senha": {
            "type": "string",
            "minLength": 6
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "senha": {
            "type": "string"
          }
        }
      },
...
    }
  }
}